backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Image processing
Pillow==10.1.0
numpy==1.26.2

# 3D model generation
trimesh==4.0.5
numpy-stl==3.1.1

# File handling
python-dotenv==1.0.0
werkzeug==3.0.1

# Utilities
marshmallow==3.20.1
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
